Implement Generic type
Generic over anything that can be encoded with MessagePack. Allows for
the use of Extension types as defined in the MessagePack spec. Provide
special methods that take Generic values, and the crate-specific
Serializer/Deserializer to allow the encoding/decoding of Extension
types into MessagePack streams. When not using those special methods,
Extension types are encoded as two-element structs instead.